+++ Changes.md
@@ -1,6 +1,15 @@
+WIP  TBD
+
+ * Restored the functionality previously intended to be provided by Readers and Writers accessors on test loaders/saver, but in a way that should actually work.
+ * Adds the CurrentReader method to TestingLoaderSaver.
+ * Adds the CurrentWriter method to TestingLoaderSaver.
+ * Adds the ReadersClosed method to TestingLoaderSaver.
+ * Adds the WritersClosed method to TestingLoaderSaver.
+
 v0.1.0  2023-08-26
 
  * Rewrote TestLoaderSaver so that loader fails before saver runs and also provides a list of Buffers once saver's writer is closed. This behaves as a better analog of the file system loader/saver as well.
+ * Adds the Buffers method to TestingLoaderSaver.
  * Added a test for the test loader/saver.
  * Improved testing of filesystem loader/saver.